Refactor game board bridge view model
