+++ docs/refactoring-task-board.md
@@ -12,7 +12,7 @@
 
 | 高     | ゲームコア | `Game` パッケージの依存関係棚卸しと公開 API の命名・アクセス制御見直し | [x] | <!-- UI からの利用方法統一を狙う重要タスク -->`GameModuleInterfaces` を導入し、UI からの GameCore 生成経路を一本化 |
 | 高     | ゲームコア | 盤面サイズや座標計算の定数を抽象化し、ユーティリティとして切り出す | [x] | <!-- 将来の盤面拡張を想定した設計 -->`BoardGeometry` ユーティリティで共通化しテストを追加 |
-| 高     | UI | SwiftUI と SpriteKit の橋渡し層に ViewModel を導入して状態管理を整理 | [ ] | <!-- 状態の単一責務化で不具合を防ぐ -->描画更新とゲーム状態を分離する |
+| 高     | UI | SwiftUI と SpriteKit の橋渡し層に ViewModel を導入して状態管理を整理 | [x] | <!-- 状態の単一責務化で不具合を防ぐ -->描画更新とゲーム状態を分離する |
 | 高     | サービス | StoreKit・Game Center・AdMob の共通プロトコル整備とモック実装作成 | [ ] | <!-- 非同期処理の標準化で信頼性を高める -->共通エラーハンドリング方針を確立 |
 | 中     | サービス | ATT・UMP の状態遷移をドキュメント化し、コードへ反映 | [ ] | <!-- 審査対応を想定した整合性確保 -->プライバシー設定画面と連携させる |
 | 中     | ドキュメント | `docs/recommended-task-list.md` を更新して負債棚卸し結果を反映 | [ ] | <!-- タスクの可視化を最新に保つ -->継続開発の羅針盤として管理 |
